Add 외래키 위반 예외처리

+++ Drawer.Application/Services/InventoryManagement/Commands/DeleteLocationCommand.cs
@@ -25,8 +25,13 @@ namespace Drawer.Application.Services.InventoryManagement.Commands
 
         public async Task<DeleteLocationResult> Handle(DeleteLocationCommand command, CancellationToken cancellationToken)
         {
+            // 위치가 참조된 곳이 있는지 확인한다.
             var location = await _locationRepository.FindByIdAsync(command.Id)
                 ?? throw new EntityNotFoundException<Location>(command.Id);
+
+            var followerExist = await _locationRepository.ExistByUpperLocationId(location.Id);
+            if (followerExist)
+                throw new AppException($"{location.Name}에 포함된 위치가 존재합니다");
 
             _locationRepository.Remove(location);
             await _locationRepository.SaveChangesAsync();
